@@ -6,7 +6,7 @@ plugins {
     id 'com.google.osdetector' version '1.6.2' apply false
     // If you want to change `org.jetbrains.kotlin.jvm` version,
     // you also need to change `org.jetbrains.kotlin:kotlin-allopen` version in `dependencies.yml`.
-    id 'org.jetbrains.kotlin.jvm' version '1.3.72' apply false
+    id 'org.jetbrains.kotlin.jvm' version '1.4.10' apply false
 
     id 'net.ltgt.errorprone' version '1.2.1' apply false
 }
@@ -207,9 +207,9 @@ allprojects {
     }
 }
 
-// Require to use JDK 13 when releasing.
+// Require to use JDK 14 when releasing.
 tasks.release.doFirst {
-    if (JavaVersion.current() != JavaVersion.VERSION_13) {
-        throw new IllegalStateException("You must release using JDK 13.");
+    if (JavaVersion.current() != JavaVersion.VERSION_14) {
+        throw new IllegalStateException("You must release using JDK 14.");
     }
 }
