@@ -42,7 +42,7 @@ import com.google.common.annotations.VisibleForTesting;
  * requests), or those who do not provision random trace ids. It is not appropriate for collectors
  * as the sampling decision isn't idempotent (consistent based on trace id).
  *
- * <h3>Implementation</h3>
+ * <h2>Implementation</h2>
  *
  * <p>This initializes a random bitset of size 100 (corresponding to 1% granularity). This means
  * that it is accurate in units of 100 traces. At runtime, this loops through the bitset, returning
