@@ -107,29 +107,32 @@ public final class ArmeriaConfigurationUtil {
 
         server.meterRegistry(meterRegistry);
 
-        if (settings.isEnableMetrics() && !Strings.isNullOrEmpty(settings.getMetricsPath())) {
-            final boolean hasPrometheus = hasAllClasses(
-                    "io.micrometer.prometheus.PrometheusMeterRegistry",
-                    "io.prometheus.client.CollectorRegistry");
+        if (settings.isEnableMetrics()) {
+            server.decorator(MetricCollectingService.newDecorator(meterIdPrefixFunction));
 
-            final boolean addedPrometheusExposition;
-            if (hasPrometheus) {
-                addedPrometheusExposition = PrometheusSupport.addExposition(settings, server, meterRegistry);
-            } else {
-                addedPrometheusExposition = false;
-            }
+            if (!Strings.isNullOrEmpty(settings.getMetricsPath())) {
+                final boolean hasPrometheus = hasAllClasses(
+                        "io.micrometer.prometheus.PrometheusMeterRegistry",
+                        "io.prometheus.client.CollectorRegistry");
 
-            if (!addedPrometheusExposition) {
-                final boolean hasDropwizard = hasAllClasses(
-                        "io.micrometer.core.instrument.dropwizard.DropwizardMeterRegistry",
-                        "com.codahale.metrics.MetricRegistry",
-                        "com.codahale.metrics.json.MetricsModule");
-                if (hasDropwizard) {
-                    DropwizardSupport.addExposition(settings, server, meterRegistry);
+                final boolean addedPrometheusExposition;
+                if (hasPrometheus) {
+                    addedPrometheusExposition =
+                            PrometheusSupport.addExposition(settings, server, meterRegistry);
+                } else {
+                    addedPrometheusExposition = false;
+                }
+
+                if (!addedPrometheusExposition) {
+                    final boolean hasDropwizard = hasAllClasses(
+                            "io.micrometer.core.instrument.dropwizard.DropwizardMeterRegistry",
+                            "com.codahale.metrics.MetricRegistry",
+                            "com.codahale.metrics.json.MetricsModule");
+                    if (hasDropwizard) {
+                        DropwizardSupport.addExposition(settings, server, meterRegistry);
+                    }
                 }
             }
-
-            server.decorator(MetricCollectingService.newDecorator(meterIdPrefixFunction));
         }
 
         if (settings.getSsl() != null) {
